requirements.txt
# Core dependencies
google-cloud-bigquery>=3.40.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# Data handling
db-dtypes>=1.5.0
pyarrow>=21.0.0

# Output formatting (optional)
tabulate>=0.9.0
colorama>=0.4.6

# Type checking (optional)
typing-extensions>=4.8.0

